creadas las entidades, habilidades,proyectos y educacion, e implementado el jwt, salvo que hay errores en el pre autorized

+++ backendargentinaprograma/backendargentinaprograma/src/main/java/com/MiguelSotelo/backendargentinaprograma/repository/PersonaRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface PersonaRepository extends JpaRepository<Persona, Long> {
 
-    void deletePersonaById(int id);
+    void deletePersonaById(Long id);
 
     Optional <Persona>findPersonaById(Long id);
 }
